fix: long type으로 변경 input 범위에 따라

+++ src/main/java/bakejoon/greedy/수들의합/controller/NumberController.java
@@ -11,7 +11,7 @@ public class NumberController {
 	}
 
 	public static void controlNumber(Scanner scanner) {
-		int sum = InputView.inputSum(scanner);
+		long sum = InputView.inputSum(scanner);
 		OutputView.printMax((NumbersFactory.makeMax(sum)));
 	}
 }
