+++ src/main/java/bakejoon/greedy/수들의합/view/OutputView.java
@@ -5,7 +5,7 @@ public class OutputView {
 	private OutputView() {
 	}
 
-	public static void printMax(int max) {
+	public static void printMax(long max) {
 		System.out.println(max);
 	}
 }
